Fix bug on CUDA fib kernel to support long int

+++ recApps/fib/src/fib.h
@@ -1,4 +1,4 @@
 #ifndef __FIB_H__
 #define __FIB_H__
-void fibGPU(int n, long long int* arrayN);
+void fibGPU(int n, long int* arrayN);
 #endif
